requirements.txt
Django
line-bot-sdk
gunicorn
pandas